Tests do not run on cassandra as arguments are passed to JVM
Fix the tests so that args are conditionally set based on cluster type.
This is a temp solution as a more correct solution would be to introduce
urchin_args that will be passed when urchin nodes/cluster are involved
and jvm_args when origin is created - this will allow setting options
for both types and using what is needed in every case.

There are cases in which jvm_args holds configuration settiongs that
should apply to urchin as well.

Signed-off-by: Shlomi Livne <shlomi@cloudius-systems.com>