Add simple_add_new_node_while_adding_info tests
Add two tests that add a new node while data is written to the cluster.
The first add with rf=1 and checks with cl=ONE that all data exists.
The second add with rf=2 and checks with cl=TWO that all data exists -
in this case repair may solvbe the issue - yet its not clear it is
needed.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>