cql_additional_tests.py: Bring back removed CQL tests
Commit 3158f8277edc28911a76deb9a74aea38c3d9ef31 removes
a bunch of CQL tests, since those were moved to the
java unittest suite. They are useful to us, so let's
bring them back in cql_additional tests, with a bunch
of cleanups and use of session instead of cursor, which
is the preferred way of naming the connections with the
database.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>