doxygen: Do not hardcode doxypy.py's path
When installing doxypy with pip, the doxypy.py entry
point might not be installed where the Doxyfile has
it hardcoded. Let's drop the hardcoded path and let
the $PATH configuration to properly resolve the location
of the script.

This solves a problem with the automated generation
of the scylla test plan in our jenkins nodes.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>