Add an iterative test for adding and decommissioning nodes
Add tests that will stress a bit the adding and decommissioning of
nodes.  The tests are split to ones that in each iteration add a single
node and then decomission a single node / tests that add multiple nodes
and then decomission multiple nodes. Following each node addition
additional data is inserted into the cluster. All tests at the end have
a single node that should hold all the data.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>