repair_additional_test: add repair_ttl_update_test
This test is similar to repair_cell_update_test, except that the updated
cell gets an expiration time (TTL), and we verify that the same cell with
the same expiration appears on the repaired node.

Sadly, the implementation is made complicated by CQL lack of an accurate
way to retrieve the expiration time of a cell (it just lets you retrieve
the time *left* until the expiration). So we resort to the much uglier
technique of reading the sstables and checking the values actually stored
in them.

Fixes #123

Signed-off-by: Nadav Har'El <nyh@scylladb.com>