repair_additional_test: test "-pr" repair option
This test tests the "-pr" (partitioner range, aka primary range) repair
option.

The test works by setting up two replicas which each has a different
set of 1000 partitions (exactly as in the repair_disjoint_data_test test).
Running "nodetool repair -pr" on one node will only repair half of the
token ranges, so the result will be more partitions in each node (close
to 1500), but not the full set of 2000. Running "nodetool repair -pr"
also on the second node will complete the repair, and bring us to
the goal of 2000 partitions on each node.

Fixes #133.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>