repair_additional_test: remove temporary file
The read_sstable() code left behind a json file in /tmp, which was
annoying. Instead, use tempfile.TemporaryFile() which guarantees
that the file is deleted when the python object is no longer is
in use (on Linux, the guarantee is even stronger because the file
is deleted immediately, so even a Python crash cannot leave it
behind).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>