TestBackupRestore: replace get_cf_name() with get_non_snapshot_cf_dir()
Currently get_cf_name() was used to find a directory of a newly created
CF. In particular this directory can't have any snapshot. Change
the logic of searching the above directory to explicitly search the CF
directory that doesn't contain a snapshot with a given tag.

The problem with the original get_cf_name() was that if CF was not filled
with any data - there won't be any sstables created.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>