repair_additional_test: kill repair participant
While the previous test checked for killing the master node of a repair in
the middle of a repair, this test complicates matters by killing a different
node which participates in the repair, as see that the master figures it
out. This test would not work on Cassandra - see
https://support.datastax.com/hc/en-us/articles/204226119-Troubleshooting-hanging-repairs
but it would be nice if it worked on Scylla.

This test "almost" works on Scylla nows - streaming can take a very long
time to give up on the dead node, but eventually it will. However, the
test eventually fails because it finds "ERROR" messages in the log.
I should fix this (to consider these "ERRORs" acceptible) in a follow-up
patch.

Refs #124.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>