repair_additional_test.py: check for graceful shutdown of repairing node
This test is a reproducer for scylla issue 699: We stop a node in the
middle of repair (for which it is the master), and would like to see
that it shuts down gracefully, without a crash.

Unfortunately, at the moment it does crash (assertion failure
on local_is_initialized()), so this test fail. This should be fixed
(this is scylla issue 699).

Refs #124.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>