repair_addition_test: make repair_kill_2_test succeed
repair_kill_2_test failed just because there were "ERROR" messages in the
log by the end of the test. But these error messages are to be expected
(they happen because of RPC failures when the repair slave goes down),
and should not fail this test. So these patch ignores these log errors.

Fixes #125

Signed-off-by: Nadav Har'El <nyh@scylladb.com>