repair_additional_test: another test for repair with down node
This is another, far more elaborate, test for issue #127, about repair
while one of the nodes is down.

In this test we check not only that repair fails if one of the replicas
is down, but we confirm that the replicas which are *not* down are
actually repaired as much as possible, so when the dead node comes back
again, if it is send a repair command to, we end up with a fully repaired
system.

Strangely, while this test passes well in Scylla, it fails disasterously
on Apache Cassandra: Some partitions appear to be lost in the repair
(we end up with slightly fewer than the expect 2000 partitions), and
we get a bunch of "LEAK" errors in the log:

ERROR [Reference-Reaper:1] 2016-01-21 13:05:34,095 Ref.java:179 - LEAK DETECTED: a reference (org.apache.cassandra.utils.concurrent.Ref$State@b5d7de7) to class org.apache.cassandra.io.sstable.SSTableReader$InstanceTidier@94016252:/home/nyh/.dtest/dtest-L52wlZ/test/node1/data/ks/cf-cf3c1cc0c02e11e5aaba3566ec3123ba/snapshots/e40a7f70-c02e-11e5-89b4-3566ec3123ba/ks-cf-ka-1 was not released before the reference was garbage collected

Signed-off-by: Nadav Har'El <nyh@scylladb.com>