migration_test: add migrate_sstable_with_schema_change_test test
Check migration of sstable with schema changed

This test is about checking that we are able to migrate content
from Cassandra that had its schema changed.

Test is about checking that we can read content properly,
then inserting a new row and checking that the full content
of the table is correct.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>