nodetool_additional_test.py: Wait before connecting to the cluster
If trying to connect to the cluster at a too early stage, the API will
not be ready and the nodetool command will fail.

This patch extend the delay before the nodetool try to run command on
the cluster.

It also adds a sleep between creating a node and starting it when done
on a new thread.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>