nodetool_additional_test: Adding the concurent drain test
This patch adds a concurent test for drain.
It also add some improvement to the way the concurent test is displied
in case of failure or in the traces.

The drain test compose of running a culster load it and call drain to
one of the node.

Stress is run during the entire test and status check nodetool commands
are been used.