nodetool_additional_test: Solve timing problem in netstats_test
In the netstats_test a new node is added to a two node cluster, when the
node start streaming a netstats command is been called.

Doing the command on the new node may result in failure if not all the
APIs are register.

This fixes the issue by calling the command on the already running
nodes.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>