Fix consistency_test.TestConsistency.data_query_digest_test
The problem is in the test we have

 node1.stop()
 debug('Updating node2')
 node2.start()

When node2 starts, since it is not a empty cluster, it will first try to
contact the other seed node (node1) which is down before fallbacking to
do the feature check in system table. While node2 is retrying to
connect to seed node (300 seconds), dtest thinks node2 fails to start.

   Traceback (most recent call last):
     File "cassandra/cluster.py", line 840, in
   cassandra.cluster.Cluster.connect (cassandra/cluster.c:11146)
       self.control_connection.connect()
     File "cassandra/cluster.py", line 2088, in
   cassandra.cluster.ControlConnection.connect (cassandra/cluster.c:36955)
       self._set_new_connection(self._reconnect_internal())
     File "cassandra/cluster.py", line 2123, in
   cassandra.cluster.ControlConnection._reconnect_internal
   (cassandra/cluster.c:37811)
       raise NoHostAvailable("Unable to connect to any servers", errors)
   NoHostAvailable: ('Unable to connect to any servers', {'127.0.0.2':
   error(111, "Tried connecting to [('127.0.0.2', 9042)]. Last error:
   Connection refused")})

To fix, restart node2 before stopping node1.

Also disable hinted_handoff_enabled to prevent data get recovered by it.