tests: Use list() to find ResultSet length
One of the changes from the python driver 2.7.2 to 3.6.0
is the fact that ResultSet now is an iterator. Therefore,
it doesn't have a __len__ property that allows you to
directly call len() on it.

Many of our tests use the old behavior of ResultSet, so
let's fix them to use list() on the ResultSet prior to
trying to evaluate its length.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>