offline_tools_test.py: Ignore more nodetool stderr patterns
We recently had the following test failure:

File "/usr/lib64/python2.7/unittest/case.py", line 367, in run
    testMethod()
  File "/data/jenkins/workspace/scylla-master-dtest/label/monster/mode/release/smp/2/urchin-dtest/offline_tools_test.py", line 46, in sstablelevelreset_test
    self.assertTrue(len(error) == 0 or "Max sstable size of" in error, error)
  File "/usr/lib64/python2.7/unittest/case.py", line 460, in assertTrue
    raise self.failureException(msg)
'Picked up JAVA_TOOL_OPTIONS: -Dfile.encoding=UTF8'

This happened because the JAVA_TOOL_OPTIONS environment
variable was set. This is of course not a real problem,
so let's improve the analysis of the standard error
stream for nodetool commands, to make it correctly
ignore messages on the test's .ignore_log_patterns
attribute.

Signed-off-by: Lucas Meneghel Rodrigues <lmr@scylladb.com>