Fix compaction_is_started_on_boot_test tests
The test code is flaky and would result from time to time with errors on
not being able to compact "Compaction found an exception when reading
sstable ... : std::runtime_error (compressed reader out of sync)".

The test extracted the list of all the sstablefiles from the ks/cf directory
and then copied them with a new generation. The issue was that if there was
more then a single  sstable generation then depding on the ordering we
may mix the files of two generations in the copy - creating sstables
that are "illegal". example if the list ended with ...1-Data.db,
2-CompressionInfo.db then the new created files would be 999910-Data.db
(from 1) and 999910-CompressionInfo.db (from 2) which ofcoure could not
work.

Fixed this by extracting the list of TOC sstable files and using only
the first - and then expancing it to all the parts (Index,Data, etc.)

In addition there was a bug in extracting the # of files before starting
the node and after starting it

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>