Fix consistency_test errors
commit 95c872609a9ab3846d7d1d9828c384204426b82d imported
ConsistencyLevel from thrift_bindings.v22.Cassandra hidding the import
from the cassandra cql done a few lines up. As such references to
ConsistencyLevel in the code has been mixed up.

To fix this remove the import and update the few places in which
ConsistencyLevel refers to the thrift version to use the full path

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>