repair_additional_test: a test for N > RF
This is another basic dtest for the repair functionality.
We set up a cluster with N=3 nodes but RF=2, so different token ranges
have different sets of replicas. The test fills the database with data
while one node is down (which is possible because RF=2), and then
brings it back up and repairs it. The repair will need to repair each
token range against the correct set of neighbors.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20170103172612.32093-1-nyh@scylladb.com>