Support running multiple tests in parallel
Detect if multiple tests are running by checking if NOSE_PROCESSES is
set.
Introduce ClusterIdAllocator that allocates IDs for test clusters.
Povide two implemntation one that uses a single cluster id (0 to align
with previous used prefix 127.0.0.X) and the second that allocates id
from a pull (0-99).

When running tests in parallel - do not use the last test indication

Signed-off-by: Shlomi Livne <shlomi@monster.cloudius-systems.com>