fix sstabledump test to support multiple sstables
Tests assume a single sstable is created. Tests invoke sstabledump on
all sstabkes and concatenate the result. Each invocation of sstabledump
generates a legal json format that is an array []. The concatenation is
a non legal format as it generate a string with multiple arrays. Fix
this by replacing a "][" marker with a comma.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>