disable cache_hit_rate_read_balancing and move disable_hintted_handoff
some tests rely on connecting to a single replica and then reading data
of that replica. cache_hit_rate_read_balancing can cause a read read
request sent toa node that is a replica to be redirected to another node
- this does not align with the tests logic. Diabeling the feature makes
sure that a node that is a replica will return the data.

some tests had hintted handoff disabled by set the configuration after
nodes have been started (that will not have an effect) - moved the
configuration prior to starting the nodes

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>