get_session: get exclusive connection by default
Current get_session() supports to contact with exclusive node,
but the queries will still be balanced to all nodes.

This patch added an argument to make it configurable.

In auth_test.py, most of condition we with to establish connect
with exclusive node, and only query on that. So let's enable
exclusive by default.

Signed-off-by: Amos Kong <amos@scylladb.com>