Account for applicable auth. permissions
As Scylla is transitioning to an access-control system focused around
roles, the security model is also undergoing improvements that mirror
those in Apache Cassandra.

One such improvement is that the system no longer allows non-applicable
permissions to be granted to a resource.

CREATE is not an applicable permission for a specific table; it is only
applicable for a keyspace or all keyspaces.

While upstream dtests have evolved `auth_test.py` and `cql_tests` and
also added tests for roles, we will delay merging these changes until
roles are fully supported in Scylla. For now, making hotfixes like this
seems reasonable.