materialized_views_test: Test dropping a view while it's building
This test drops a view while it's building, and expects it to be
erased from the system tables. The test is disabled, because loading
the database with enough data that the view build process takes more
time than schema agreement is expensive and makes the test take long.
There is a unit test that performs the same validation, and this one
can be executed manually on need. We expect to re-enable it when we
have a framework with which to instrument scylla code at runtime.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>