Add dtest for per-query read repair decision
Test for scylla commit eee9bda85b0289ca9ddbe3381596e42d69b7b0db.
Set dclocal_read_repair_chance to 0.5 and check that all pages stick to
the read repair decision made on the first page. This can be validated
by checking that there are no querier_cache_drops and
querier_cache_misses during the query and that a replica either has
query_cache_lookups that is equivalent to the number of pages or none at
all. Since the query we execute is CL=ONE we should send read
requests for all pages to either one or two replicas consistently.