materialized_views_test: Fix false positives
Some tests that delegate to
mv_populating_from_existing_data_during_changes_test are expected to
fail. If a test is expected to fail but doesn't, we throw an
exception. That exception was then being caught and ignored, because
we expected a failure. This patch fixes this logic.

Also,a change introduced in 7de049bb0a4b705010c71127c32fbc832c385fc3
replaced an exception with an AssertionError, thereby
losing the original type (probably also an AssertionError) and, most
importantly, the original call stack. This patch also fixes that.

Signed-off-by: Duarte Nunes <duarte@scylladb.com>