Remove `counter_auth_test`
This test does not test anything related to counters. Instead, it tests
the authorization of access to resources like tables and keyspaces which
just so happen to contain counters. Tests for the access-control system
are better suited elsewhere.

Furthermore, Scylla now grants users default permissions on
the objects they create. Therefore, many of the assertions in this test
are outdated.