Optionally include superuser perms. in tests
The `all_authorization_operations` test doesn't account for the default
permissions granted to the `cassandra` user when it creates the objects
used for the test.

Rather than change a bunch of assertions, we remove by default
permissions for `cassandra` unless the caller indicates otherwise.

This fixes `all_authorization_operations_test`.