Correct and update "transitional" auth. tests on `master` (#609)
* Optionally wait for the default superuser

Some auth. tests depend on the background task which creates the default
superuser successfully finishing, even if auth. itself is not enabled at
this point in the test.

This change should not impact any existing test, due to the default
parameter value.

* Correct logic and update `transitional_auth_from_default_test`

This test needed to be updated to reflect the new auth. support in
Scylla, and also there were several logical errors.

- We need to wait for the default superuser to be created before we shut
  down the cluster.
- Users who login with bad credentials will be anonymous but with a
  predefined set of permissions
- These anonymous users cannot execute any access-control-modifying
  actions
- There was a typo in the last assertion