Fix `transitional_auth_betweenness_from_default_test`
When access control is not enabled, the default superuser role is
created but its authentication record is not. The record is created,
however, when the node switches to the transitional authenticator and
authorizer.

The test did previously waited for the role to be created, but not the
authentication record. Changing the test to wait for the authentication
record ensures that `cassandra` can log in as expected.