Skip `only_superuser_can_create_users_test`
Now that Scylla has a revamped access-control system, its security model
has changed.

This test should be skipped (just like it is skipped in Apache
Cassandra) since it is no longer applicable.

The comment about "NOSUPERUSER" is invalid. That the test was originally
not skipped has nothing to do with "NOSUPERUSER".