update_cluster_layout_tests: Fix simple_kill_node_while_decommissioning_test
The test does:

Force kill node2
Start node2, immediately.

It is possible that node2 restarts before node1 and node3 even nodtice
that node2 is DOWN. So when node2 restarts,
node2.start(wait_other_notice=True) will fail, because node1 and node3
won't have "gossip - InetAddress 127.0.0.2 is now UP" log

In addition, insert more data in the ks, otherwise, it is possible that
when node2 is killed, the decommission is finished already.