replace_address_test: expect ReadFailure error in addition to Unavailable and ReadTimeout
When a node is down, we can get an Unavailable exception (if the failure
detector detected the failure even before we initiated the query) or a
ReadTimeout (if it did not, and the read timed out). However, starting with
Scylla 2.3 we can also see a ReadFailure exception - if the failure detector
did not detect the node down condition, but RPC detected that the connection
was broken and therefore no response will arrive. This allows an early error
response instead of waiting for the timeout.

Adjust the test to expect the ReadFailure exception.