Repair tests (#671)
* repair_additional_test: test combination of "-pr" with -dc/-host/-local

This is a comprehensive test for combining the nodetool repair's "-pr"
option (primary range) with the options restricting the nodes involved
in the repair: -dc, -host, -local.

First of all, we verify that combining -pr with "-host" and "-dc" is
not allowed at all (there is actually a check in nodetool itself which
prohibits it, no matter what parameter is passed to "-dc").

However, "-local" *is* allowed together with "-pr", and it's not enough
that we repair only inside the DC, we also need to define the primary
ranges differently, so that every token has a primary owner inside every
DC (and not just across the entire cluster). So here we build a test
with multiple data centers and do a repair with "-pr -local" on just
the nodes of one of the data centers, and verify that this repairs the
entire token range. This reproduces (and verifies the fix) for issue 3557.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

* repair_additional_test: test "-pr" repair with multi-dc cluster

The previous patch added a test for the combination of "-pr" and "-local"
repair on a multi-dc cluster. This test checks "-pr" alone, without "-local".
This requires that a repair be done to *all* nodes in the cluster (across
all data centers), otherwise we will miss some of the token ranges.

This test is very similar to a test we already had for "-pr", but
adds the complication that there are multiple data centers. This shouldn't
make any difference in the outcome, and indeed it doesn't, and the
test works without any changes to Scylla code.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

* repair_additional_test: replace duplicate code by loops

replace duplicated code in _repair_option_pr_multi_dc_test by loops, as
requested in review. Also drop unnecessary flush() before stop()ing a node.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

* repair_additional_test: use count(*)

Use SELECT count(*) instead of SELECT * and counting the result.
Suggested by Roy during code review.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

* repair_additional_test: kill the other nodes, as intended

In repair_option_pr_dc_host_test() we intended to kill all the nodes except
one, but accidentally left 4 additional nodes alive. This did not break the
test, but better do as intended (and commented). Discovered during code review.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>