Fix simple_kill_new_node_while_bootstrapping_with_parallel_writes_in_multidc_test
When the streaming for bootstrap happens, nothing prevents when we check
the node tool status the new node has already finished joining the
cluster, so we need to wait for UN or UJ for the new node.

After the new node is killed, the other node can either see the new node
has joined the cluster but it is down or the node does not join
successfully so other nodes forget about it.