force partitioner tests to not run in parallel (#715)
cassandra cluster data was created with 127.0.0.x. Allowing the tests to
run in parallel could allocate the tests a prefix 127.0.Y and as such
could have the partitioner tests node "connect" to other parallel
clusters. Forcing the tests to run on their own insures the nodes will
use 127.0.0.x prefix.

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>