interrupt_build_process_with_resharding_half_to_max_test read timeout
The resharding tests are run a lot of threads, and on overcommitted test
systems, they can become very slow. We try to count 200,000 rows, and
by default time out after 10 seconds. In some setups, this is never enough.
In our Jenkins setup, it's sometimes enough.

This patch increases the timeout to 100 seconds, which appears to be
enough to make this test succeed on setups where it previously failed.

Fixes #740.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>