save rows list for multiple iterations
since the return type from execute (ResultSet) is an iterator, it
can be looped through only once. This led to the assert messages
being blank.In order to print the actual results when the assert fails,
a reference to the list created from the ResultSet should be saved and
then used twice in the assert body, once for assertion and once for
printing the assertion message.