Don't use node.status directly
In a couple of utility functions, we checked for node liveness by
checking if node.status == 'UP'. This is bad, since this field
doesn't get magically updated when it's read... Rather, we have
a function node.is_live() which first updates the status and then
checks if it is 'UP'. We should use this function.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>