Fix wait_for_view() when node is killed with wait_other_notice=false
The intention of the wait_for_view() code was to wait until the number of
nodes which built views is equal to the number of living nodes. However,
the implementation only counted the living nodes once, when the function
starts, so if we call wait_for_view() before we realize that one of the
nodes is dead, the wait may never finished.

We saw this problem in secondary_indexes_test.py's
test_stop_node_during_index_build which frequently failed. This test stops
a node gently, but with wait_other_notice=false, so when wait_for_view()
is called, depending on the timing, it may not yet know the node is down.

The wait_for_view() algorithm still isn't perfect. Counting the built
nodes is the wrong thing to do, because it is possible that a dead
node has finished building the view before it was killed. This will
be improved in a later patch, but this patch already solves the
problems we saw in actual tests.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>