secondary_indexes_test: mark test no longer flaky
The test_stop_node_during_index_build was flaky (sporadically failed)
because of a bug in wait_for_view() which we now fixed. So the test
should no longer be flaky, and should not be flagged as such (the
consequence of this flag was trying to run the test several more times
when it failed).

Signed-off-by: Nadav Har'El <nyh@scylladb.com>