Reimplement wait_for_view()
The wait_for_view() function is supposed to wait for the building of
a given view to have finished, but ONLY on living nodes. This was
not actually documented so we begin the patch with a comment saying
that.

But more importantly, the implementation wasn't strictly correct.
What the implementation did was to count the number of living nodes,
and if it was K, we wait until the distributed view-build status table
says that K nodes have finished. This is wrong when one of the dead
nodes was lucky enough to have finished building before it was killed,
so we will not wait for one of the live nodes to finish building.

So this patch includes a new implementation, which gets the ids of
the nodes which finished building, and confirms that all the currently
live nodes are on this list.

We do not know of an actual dtest failure that this patch will fix,
but it should prevent problems in the future.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>