secondary_indexes_test: mark test no longer flaky
The test_remove_node_during_index_build was marked flaky, probably
because of a bug in because of a bug in wait_for_view() when some of the
nodes are down, which we now fixed. So the test should hopefully no longer
be flaky, and should not be flagged as such.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>