Fix mv_populating_from_existing_data_during_node_decommission_test (#767)
* Fix mv_populating_from_existing_data_during_node_decommission_test

The wait_for_view() code failed for a node which this test *decomissioned*.
Decomissioning the node involves just a "nodetool decommission", but the
Python code still thinks this node exists in the cluster, and (at least for
a while), alive. But when we try to call node.hostid(), we fail with an
exception. We need to ignore this exception, and just considered this node
to be dead - rather than fail the test.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

* wait_for_view: be more specific in the kind of exception expected

When node.hostid() fails with a NodetoolError, we know it is because
there is a problem with the node (e.g., decomissioned) and can consider
it dead. In case of other exceptions, we don't know what happened -
better just pass the exception through and fail the test.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>