materialized_views_test: increase cluster stop timeout
As explained in issues #4019 and #3966, shutdown may hang for up 5 minutes
(the new timeout set in storage_proxy::send_to_endpoint()) while a view
build step is stuck trying to communicate with another node we previously
killed.

But cluster.stop()'s timeout was hard-coded to 127 seconds (2 minutes
and 7 seconds) which wasn't enough. In this patch we use a new optional
parameter (added to CCM in a separate patch) to choose a higher timeout,
of 6 minutes.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>