rebuild_test.py: fix flaky simple_rebuild_test
The simple_rebuild_test tries to run "nodetool rebuild" twice concurrently,
and verify that only one of them succeeds, and the second fails.

The existing code (written by the Cassandra guys) has duplicate code
for running the two rebuilds. Each of these duplicates can fail; Usually
the second one fails but sometimes (when the timing is just right), the
first one will fail. All we care is that exactly one of them succeeded.

But the code which tested for failure was different in both cases! The
second code, which more commonly failed, it just tested for any failure.
But the first copy of the code, which more rarely failed, tried to look for
a specific error message. The problem is that they looked for the wrong
error message. The real error message we get is:

    nodetool: Scylla API server HTTP POST to URL '/storage_service/rebuild' failed: Operation rebuild is in progress, try again

But that is not what was tested...

The fix is make both checks the same, and check for the up-to-date error
message. If we get a *different* error message in either of the rebuild
runs, the test is failed as well.

Fixes #810.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>