materialized_views_test.py: add virtual_columns_schema_test
This adds a reproducer for https://github.com/scylladb/scylla/issues/4339
That issue is about the view_virtual_columns table not being synced between
nodes. The test creates a view which has virtual columns on one node while a
second node is down, and then brings up the second node and checks if
it received the view_virtual_columns table. Before the patch for 4339
this test failed, after it, it succeeds.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>