fix TestCountersOnMultipleNodes.counter_consistency_node_repair_test
update _verify_data_repair to wait for cql port to be open after
stopping and starting node. This function is called before repair is
started and since bcql port was not open repair would fail.
update populate_data to use consistency all so that we are sure the data
is fully written to all replicas before we start doing steps in any test

Signed-off-by: Shlomi Livne <shlomi@scylladb.com>