Don't ignore the timezone when parsing sstabledump output
I can't figure out if there is something different on my machine, but
here is what I observe:

If I add both a timestamp and a map<timestamp, text> to a table, cqlsh
displays both in a consistent way:

2005-07-14 19:30:00.000000+0000 | {'2005-07-14 19:30:00.000000+0000': 'a'}

But sstabledump uses UTC for one but not the other:

{ "name" : "k", "value" : "2005-07-14 19:30:00.000Z" }
{ "name" : "r", "path" : [ "2005-07-14 12:30-0700" ], "value" : "a" }

The python code reading the output replaces the timezone with
utc. That is a nop for the first entry, but not the second.

With this change I am able to, with a scylla from master, run

nosetests -v  user_types_test.py sstabledump_test.py cqlsh_tests/cqlsh_tests.py

and all tests pass. Without it, I get a failure in sstabledump_all_datatypes_test:

AssertionError: {2005-07-14 12:30:00+0000: '1', 2005-07-14 13:30:00+0000: '2'} != {datetime.datetime(2005, 7, 14, 6, 30, tzinfo=tzutc()): '2', datetime.datetime(2 [truncated]...
- {2005-07-14 12:30:00+0000: '1', 2005-07-14 13:30:00+0000: '2'}
+ {datetime.datetime(2005, 7, 14, 5, 30, tzinfo=tzutc()): '1',
+  datetime.datetime(2005, 7, 14, 6, 30, tzinfo=tzutc()): '2'}

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>