dtest: introduce RandomClusterIdAllocator
Add support for an optional environment variable: CLUSTER_ID_ALLOCATOR.
If not set, then automatically select either the RandomClusterIdAllocator
or the SingleClusterIdAllocator based on NOSE_PROCESSES > 0.

The RandomClusterIdAllocator draws random IDs from range(1, 255)
and tries to allocate them by creating a symbolic link in the shared directory
($HOME/.dtest), pointing at the cluster base directory.

If the link happens to exist, just retry with a different random ID
for up to 10 times.  In the future the cluster the symbolic link points
to can be examined to determine if it's still running or already dead
(maybe from a previous run) and then its ID can be safely recycled.

This should fix the error storms we see with dtest-release. Our latest analysis is that:

1. one of the tests times out (on the 7200 seconds timeout)
2. nosetests restarts the worker
3. the new worker instantiates a MultiProcessClusterIdAllocator
4. this causes it to go out of sync with the other workers and when
   it starts stepping on their IDs, our cluster cleanup code just
   kills the other clusters - causing the error storms we see.

This change uses the filesystem for synchronizing ID allocation
without the need for shared memory and will work even if unrelated
nosetests run unintentionally on the same machine.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>