dtest: prevent termination of worker during timeout handling
We see that nosetests multiprocess pluging has a short, 5 seconds timeout
for the timeout handler to cleanup.  We tried overriding it, but we can't
get to the internal MultiProcessTestRunner attribute.

To prevent this timeout, we need to cleanup as quickly as possible
so do the minimum, leaving the test directory as is and just
kill the node processes in _force_clean.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>