simple_bootstrap_test: fix nodes total_size assertion
Introduced in e5ad7e5620cd7564ae167cfa421a136310bfc42f.
This looks like a typo because it makes no sense to assert that
node1's total_size is equal to itself.
I guess the intention was to assert that the total size on both
nodes is roughly the same.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>