nested_user_types_test: fix test flakiness
The test uses a configuration that in combination with
the test queries sequence can cause flakiness.
The configuration is a 3 node cluster with RF=2, combined
with using node1 as a coordinator to write a record
to a random partition and imediately reading back for
verification using in both, the write and the read
the default CL which is ONE.
if the record is written to node2 with replica on node3,
and the background write to node3 is delayed or dropped,
and in addition the CL=1 read is done from node3, the just
written data will not be present in the query result since
it is not yet on node3 and the test will fail even though
the data is present in the cluster.
My assumption is that the CL=ONE was not taken into account
while writing the test, and the fix is to change the default
CL for this tests to LOCAL_QUORUM.

I also didn't understand the RF=2 configuration, it is an
odd configuration that doesn't seam to satifay any specific
requirement, I think that tests should in the lack of a specific
requirement should use RF=1 for single  node tests and RF=3
for multi node tests as as default (with at least 3 nodes of
course).

Fixes #1025