commitlog_test, nodetool_additional_test: adjust expected error messages to libstdc++fs 9
std::filesystem_error::what() changed in version 9 of the standard c++ library
and it does not include the textual system error description as it used to.

It might be considered as a bug in the library but we can simply make
the dtest more permissive about it and check only the error message prefix.

For example, the following c++ code generates different outputs with g++-8 vs. g++-9:
    #include <iostream>
    #include <system_error>
    #include <filesystem>

    int main()
    {
        int e = 13;
        auto ec = std::error_code(e, std::system_category());
        auto fse = std::filesystem::filesystem_error("op failed", ec);

        std::cout << "filesystem_error: " << fse.what() << '\n';
    }

With g++-8:
    filesystem_error: filesystem error: op failed: Permission denied

With g++-9:
    filesystem_error: filesystem error: op failed

Fixes #1035

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>